requirements.txt
blinker==1.9.0
click==8.3.1
dlib==19.24.1
face-recognition==1.3.0
face_recognition_models==0.3.0
Flask==3.1.3
Flask-Login==0.6.3
gunicorn==25.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
numpy==1.26.4
opencv-python-headless==4.13.0.92
packaging==26.0
pandas==2.3.3
pillow==12.1.1
python-dateutil==2.9.0.post0
pytz==2026.1.post1
six==1.17.0
Werkzeug==3.1.6